Break the App's API controller to multiple controller classes.
