+++ src/routes.php
@@ -1,36 +1,42 @@
 <?php
 // Routes
 
-$app->get('/', 'controllers/home:index')->setName('home')->add($middleware['view']);
+// Home
+$app->get('/', 'controllers/app:home')->setName('home')->add($middleware['view']);
 
 // Authentication
 $app->get('/auth/login', 'controllers/auth:loginPage')->setName('auth/login')->add($middleware['view']);
 $app->post('/auth/login', 'controllers/auth:loginProcess');
 $app->any('/auth/logout', 'controllers/auth:logoutProcess');
 
-// Index
-$app->get('/app', 'controllers/app:index')->setName('app/index')->add($middleware['view']);
+// App page
+$app->get('/app', 'controllers/app:app')->setName('app/index')->add($middleware['view']);
 
 // API routes
 $app->group('/api', function () use ($app, $middleware) {
-    $app->get('/balance[/{accountUID}[/{date}]]', 'controllers/app:getBalance');
-    $app->get('/account/{UID}', 'controllers/app:getAccount');
-    $app->get('/accounts', 'controllers/app:getAccounts');
-    $app->get('/transaction/{UID}', 'controllers/app:getTransaction');
-    $app->get('/transactions/tagged/{tags}', 'controllers/app:getTransactionsTagged');
-    $app->get('/transactions[/{accountUID}[/{year}[/{month}[/{date}]]]]', 'controllers/app:getTransactions');
+    // Accounts
+    $app->get('/account/{UID}', 'controllers/api/accounts:getOne');
+    $app->get('/accounts', 'controllers/api/accounts:getAll');
+    $app->post('/accounts', 'controllers/api/accounts:post')->add($middleware['csrfp']);
+    $app->put('/account/{UID}', 'controllers/api/accounts:put')->add($middleware['csrfp']);
+    $app->delete('/account/{UID}', 'controllers/api/accounts:delete')->add($middleware['csrfp']);
 
-    $app->get('/stats/expense[/{year}[/{month}[/{date}]]]', 'controllers/app:getExpense');
-    $app->get('/stats/income[/{year}[/{month}[/{date}]]]', 'controllers/app:getIncome');
+    // Transactions
+    $app->get('/transaction/{UID}', 'controllers/api/transactions:getOne');
+    $app->get('/transactions/tagged/{tags}', 'controllers/api/transactions:getSomeTagged');
+    $app->get('/transactions[/{accountUID}[/{year}[/{month}[/{date}]]]]', 'controllers/api/transactions:getSome');
+    $app->post('/transactions', 'controllers/api/transactions:post')->add($middleware['csrfp']);
+    $app->put('/transaction/{UID}', 'controllers/api/transactions:put')->add($middleware['csrfp']);
+    $app->delete('/transaction/{UID}', 'controllers/api/transactions:delete')->add($middleware['csrfp']);
 
-    $app->post('/accounts', 'controllers/app:postNewAccount')->add($middleware['csrfp']);
-    $app->put('/account/{UID}', 'controllers/app:putAccount')->add($middleware['csrfp']);
-    $app->delete('/account/{UID}', 'controllers/app:deleteAccount')->add($middleware['csrfp']);
-    $app->post('/transactions', 'controllers/app:postNewTransaction')->add($middleware['csrfp']);
-    $app->put('/transaction/{UID}', 'controllers/app:putTransaction')->add($middleware['csrfp']);
-    $app->delete('/transaction/{UID}', 'controllers/app:deleteTransaction')->add($middleware['csrfp']);
+    // Stats
+    $app->get('/balance[/{accountUID}[/{date}]]', 'controllers/api/stats:getBalance');
+    $app->get('/stats/balance[/{accountUID}[/{date}]]', 'controllers/api/stats:getBalance');
+    $app->get('/stats/expense[/{year}[/{month}[/{date}]]]', 'controllers/api/stats:getExpense');
+    $app->get('/stats/income[/{year}[/{month}[/{date}]]]', 'controllers/api/stats:getIncome');
 
-    $app->post('/profile', 'controllers/app:editProfile')->add($middleware['csrfp']);
+    // User-related
+    $app->patch('/profile', 'controllers/api/users:patchProfile')->add($middleware['csrfp']);
 });
 
 // User-related
